Refactor pages directory structure
