+++ src/Wrapper/Router.tsx
@@ -4,23 +4,20 @@ import {Switch, Route, BrowserRouter, Redirect} from 'react-router-dom';
 
 import {stateTypes} from 'utilities';
 
-import LoginForm from 'pages/general/accountForms/LoginForm';
-import RegisterForm from 'pages/general/accountForms/RegisterForm';
-import RequestPasswordForm from 'pages/general/accountForms/RequestPasswordForm';
-import SetPasswordForm from 'pages/general/accountForms/SetPasswordForm';
-import VerifyForm from 'pages/general/accountForms/VerifyForm';
-import VerifyWall from 'pages/general/accountForms/VerifyWall';
-import FormPage from 'pages/general/FormPage';
-import NotFoundPage from 'pages/general/NotFoundPage';
-import DashBoardPage from 'pages/dashboard/DashboardPage';
+import {FormPage, NotFoundPage, DashboardPage} from 'pages';
+import {
+    LoginForm,
+    RegisterForm,
+    RequestPasswordForm,
+    SetPasswordForm,
+    VerifyForm,
+    VerifyWall,
+} from 'pages/authentication';
+import {ConfigList, ConfigEditor} from 'pages/configuration';
 
-import LoginImage from 'assets/images/secure.svg';
-import VerifyImage from 'assets/images/letter.svg';
+import {LoaderOverlay, MessageQueue} from 'components';
 
-import LoaderOverlay from 'components/overlays/LoaderOverlay';
-import MessageQueue from 'components/messages/MessageQueue';
-import ConfigList from 'pages/dashboard/configsTab/list/ConfigList';
-import ConfigEditorWrapper from 'pages/dashboard/configsTab/editor/ConfigEditorWrapper';
+import {SecureImage, LetterImage} from 'assets';
 
 interface RouterProps {
     loggingIn: boolean;
@@ -41,12 +38,12 @@ function Router(props: RouterProps) {
                         {!props.loggingIn && props.loggedIn && (
                             <React.Fragment>
                                 {props.account?.email_verified !== true && (
-                                    <FormPage image={VerifyImage}>
+                                    <FormPage image={LetterImage}>
                                         <VerifyWall />
                                     </FormPage>
                                 )}
                                 {props.account?.email_verified && (
-                                    <DashBoardPage>
+                                    <DashboardPage>
                                         <Switch>
                                             <Route exact path='/configurations'>
                                                 <ConfigList />
@@ -56,7 +53,7 @@ function Router(props: RouterProps) {
                                                 path='/configuration/:survey_name'
                                             >
                                                 <ConfigList />
-                                                <ConfigEditorWrapper />
+                                                <ConfigEditor />
                                             </Route>
                                             <Route exact path='/results'>
                                                 <h3>Results</h3>
@@ -65,7 +62,7 @@ function Router(props: RouterProps) {
                                                 <h3>Account</h3>
                                             </Route>
                                         </Switch>
-                                    </DashBoardPage>
+                                    </DashboardPage>
                                 )}
                             </React.Fragment>
                         )}
@@ -75,7 +72,7 @@ function Router(props: RouterProps) {
                     </Route>
                     <Route path='(/login|/register)'>
                         {!props.loggingIn && !props.loggedIn && (
-                            <FormPage image={LoginImage}>
+                            <FormPage image={SecureImage}>
                                 <Switch>
                                     <Route exact path='/login'>
                                         <LoginForm />
@@ -91,12 +88,12 @@ function Router(props: RouterProps) {
                         )}
                     </Route>
                     <Route path='/verify'>
-                        <FormPage image={VerifyImage}>
+                        <FormPage image={LetterImage}>
                             <VerifyForm />
                         </FormPage>
                     </Route>
                     <Route path='(/request-password|/set-password)'>
-                        <FormPage image={LoginImage}>
+                        <FormPage image={SecureImage}>
                             <Switch>
                                 <Route exact path='/request-password'>
                                     <RequestPasswordForm />
