+++ src/assets/index.ts
@@ -1,1 +1,9 @@
 export {default as icons} from './icons/icons';
+
+export {default as EmptyImage} from './images/empty.svg';
+export {default as LetterImage} from './images/letter.svg';
+export {default as SecureImage} from './images/secure.svg';
+export {default as SurveyImage} from './images/survey.svg';
+
+export {default as RocketLogo} from './branding/rocket.svg';
+export {default as RocketLogoLight} from './branding/rocket-light.svg';
