+++ src/components/messages/MessageComponent.tsx
@@ -1,6 +1,6 @@
 import React, {useEffect, useState} from 'react';
 
-import stateTypes from 'utilities/types/state-types';
+import {stateTypes} from 'utilities';
 
 import {icons} from 'assets';
