+++ src/components/navbar/components/NavbarContent.tsx
@@ -2,11 +2,11 @@ import React from 'react';
 import {useHistory, useLocation} from 'react-router-dom';
 import {connect} from 'react-redux';
 
-import stateTypes from 'utilities/types/state-types';
-import dispatcher from 'utilities/dispatcher';
+import {stateTypes, dispatcher} from 'utilities';
 
 import NavbarButton from './NavbarButton';
 import NavbarLogo from './NavbarLogo';
+
 import {icons} from 'assets';
 
 interface NavbarContentProps {
