+++ src/pages/authentication/index.ts
@@ -1,0 +1,6 @@
+export {default as LoginForm} from './LoginForm';
+export {default as RegisterForm} from './RegisterForm';
+export {default as RequestPasswordForm} from './RequestPasswordForm';
+export {default as SetPasswordForm} from './SetPasswordForm';
+export {default as VerifyForm} from './VerifyForm';
+export {default as VerifyWall} from './VerifyWall';
