+++ src/pages/configuration/editor/control-strip/control-strip.tsx
