+++ src/pages/configuration/editor/control-strip/visual-button.tsx
