+++ src/pages/configuration/editor/control-strip/visual-draft-strip.tsx
