+++ src/pages/configuration/editor/control-strip/visual-published-strip.tsx
