+++ src/pages/configuration/editor/control-strip/visual-strip.tsx
