+++ src/pages/configuration/editor/fieldConfig/SelectionFieldConfigForm.tsx
