+++ src/pages/configuration/index.ts
@@ -1,0 +1,2 @@
+export {default as ConfigEditor} from './editor/ConfigEditorWrapper';
+export {default as ConfigList} from './list/ConfigList';
