+++ src/pages/configuration/list/ConfigPreviewPanel.tsx
