+++ src/pages/index.ts
@@ -1,0 +1,3 @@
+export {default as FormPage} from './FormPage';
+export {default as NotFoundPage} from './NotFoundPage';
+export {default as DashboardPage} from './DashboardPage';
